+++ flutter-hms-analytics/android/src/main/java/com/huawei/hms/flutter/analytics/logger/HMSLogger.java
@@ -38,7 +38,7 @@ public final class HMSLogger {
 
     private static final String KIT = "Analytics";
     private static final String PLATFORM = "Flutter";
-    private static final String VERSION = "6.5.0.300";
+    private static final String VERSION = "6.5.0.301";
 
     private static final String SERVICE = "Cross-Platform";