+++ flutter-hms-analytics/ios/Classes/Logger.swift
@@ -19,9 +19,8 @@ import SystemConfiguration
 import CoreTelephony
 import Foundation
 
-let Log = Logger(version: "5.1.0.301", service: "FlutterAnalytics")
+let Log = Logger(version: "6.5.0.301", service: "FlutterAnalytics")
 
-/// Custom Logger Class to listen api log events from RN Side.
 final class Logger {
 
   /// Info Types to declare success and failure scenarios.