+++ flutter-hms-analytics/CHANGELOG.md
@@ -1,10 +1,15 @@
+## 6.5.0+301
+
+- **Breaking Change:** Added the `getInstance({String routePolicy})` method. With this version `getInstance` method must be called from the application to initialize plugin in iOS platform.
+- Support the `setMinActivitySessions`, `setCollectAdsIdEnabled` and `addDefaultEventParams` methods on iOS devices.
+
 ## 6.5.0+300
 
-- **[Breaking Change]** Added the ```getInstance({String routePolicy})```  method. With this version ```getInstance``` method must be called from the application to initialize plugin in Android platform.
-- Added events to [HAEventType](lib/src/constants/analytics_constants.dart). 
+- **[Breaking Change]** Added the `getInstance({String routePolicy})` method. With this version `getInstance` method must be called from the application to initialize plugin in Android platform.
+- Added events to [HAEventType](lib/src/constants/analytics_constants.dart).
 - Added parameters to [HAParamType](lib/src/constants/analytics_constants.dart).
 - Added attributes to [HAUserProfileType](lib/src/constants/analytics_constants.dart).
-- Added ```setChannel```, ```setPropertyCollection``` and ```setCustomReferrer``` methods to [HMSAnalytics](lib/src/hms_analytics.dart) class.
+- Added `setChannel`, `setPropertyCollection` and `setCustomReferrer` methods to [HMSAnalytics](lib/src/hms_analytics.dart) class.
 - Supported on devices running Android 12.
 
 ## 6.2.0+302